[REFACTOR] Interactive parsing
Code handling the interactive parsing put in another module.

Modified files :
- standard_syntax.py
- ../../utils/interactive_parsing.py
